Update RecordWho plugin to be compatible with Errbot v5.1.3
At some point a `name` keyword argument was added to the BotPlugin
`__init__()` function, so for future compatibility we will just pass all
arguments when invoking `super()`.

The plug file may not need to be updated, but it confirms to the inital
plug generated by Errbot.

Some of the other changes may not be necessary, but they clean things up
a bit.