Identify with nickserv
The #linux.conf.au channel now requires users to be identified with
nickserv before they can join the channel.

However there seems to be a race condition where the bot will attempt to
join channels before it has identified with the nickserv. This prevents
the bot from join channels and collecting status without manual
intervention. A work-around has be documented in the README.

See: issue #3